Rename package texifyidea.window to texifyidea.ui, remove semicolons
