+++ src/nl/rubensten/texifyidea/ui/PreviewForm.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="nl.rubensten.texifyidea.window.PreviewForm">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="nl.rubensten.texifyidea.ui.PreviewForm">
   <grid id="27dc6" binding="panel" default-binding="true" layout-manager="GridLayoutManager" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -17,11 +17,10 @@
         <properties/>
         <border type="none"/>
         <children>
-          <component id="f114e" class="nl.rubensten.texifyidea.window.ImagePanel" binding="previewPanel">
+          <component id="7ddc2" class="nl.rubensten.texifyidea.ui.ImagePanel" binding="previewPanel">
             <constraints>
               <splitpane position="left"/>
             </constraints>
-            <properties/>
           </component>
           <grid id="43813" layout-manager="GridLayoutManager" row-count="4" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
             <margin top="0" left="0" bottom="0" right="0"/>
